componentização do menu.
