+++ src/Components/Home/HomeStyle.js
@@ -77,7 +77,10 @@ display: "flex",
 overflow: "auto",
 flexDirection: "column",
 border: "1px solid gray"
-}
+},
+fixedHeight: {
+    height: 445
+    }
 }));
 
 export default useStyles;
