+++ src/Components/Menu/MenuAdmin.js
@@ -1,0 +1,88 @@
+import React, { useState } from "react";
+import Drawer from "@material-ui/core/Drawer";
+import clsx from "clsx";
+import { useStyles } from "../Home/HomeStyle";
+import List from "@material-ui/core/List";
+import Divider from "@material-ui/core/Divider";
+import { mainListItems, secondaryListItems } from "../Menu/ListItems";
+import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
+import IconButton from "@material-ui/core/IconButton";
+
+import Logo from "../Logo/Logo";
+import NotificationsIcon from "@material-ui/icons/Notifications";
+//import Link from "@material-ui/core/Link";
+import MenuIcon from "@material-ui/icons/Menu";
+import AppBar from "@material-ui/core/AppBar";
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
+import Badge from "@material-ui/core/Badge";
+
+
+
+export const MenuAdmin = () => {
+    const classes = useStyles();
+    const [open, setOpen] = useState(true);
+        const handleDrawerOpen = () => {
+        setOpen(true);
+         };
+        const handleDrawerClose = () => {
+        setOpen(false);
+        };
+
+    return (
+        <>
+        <AppBar
+        position="absolute"
+        className={clsx(classes.appBar, open && classes.appBarShift)}
+      >
+        <Toolbar className={classes.toolbar}>
+          <IconButton
+            edge="start"
+            color="inherit"
+            aria-label="open drawer"
+            onClick={handleDrawerOpen}
+            className={clsx(
+              classes.menuButton,
+              open && classes.menuButtonHidden
+            )}
+          >
+            <MenuIcon />
+          </IconButton>
+          <Typography
+            component="h1"
+            variant="h6"
+            color="inherit"
+            noWrap
+            className={classes.title}
+          >
+            <Logo />
+            Dashboard - DataCov
+          </Typography>
+          <IconButton color="inherit">
+            <Badge badgeContent={4} color="secondary">
+              <NotificationsIcon />
+            </Badge>
+          </IconButton>
+        </Toolbar>
+      </AppBar>
+        <Drawer
+        variant="permanent"
+        classes={{
+          paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
+        }}
+        open={open}
+      >
+        <div className={classes.toolbarIcon}>
+          <IconButton onClick={handleDrawerClose}>
+            <ChevronLeftIcon />
+          </IconButton>
+        </div>
+        <Divider />
+        <List>{mainListItems}</List>
+        <Divider />
+        <List>{secondaryListItems}</List>
+      </Drawer>
+      </>
+    )
+}
+export default MenuAdmin;
